Fix: 테스트시 세션 검색이 되지 않는 문제 수정

+++ session/src/main/java/com/sparta/moim/session/session/application/dto/result/SearchSessionResult.java
@@ -11,7 +11,7 @@ public record SearchSessionResult(
     int page,
     int content
 ) {
-  public static SearchSessionResult search(List<Session> sessions, int total, int page, int content) {
+  public static SearchSessionResult search(List<Session> sessions, long total, int page, int content) {
     return SearchSessionResult.builder()
         .sessions(
             sessions.stream().map(session ->
